+++ web/src/containers/SessionEndingAlert.js
@@ -1,4 +1,4 @@
-import { Alert, Button } from '@cmsgov/design-system';
+import { Dialog, Button } from '@cmsgov/design-system';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
@@ -12,6 +12,7 @@ import { extendSession, logout } from '../actions/auth';
 const SessionEndingAlert = ({
   isSessionEnding,
   isExtendingSession,
+  isLoggingOut,
   expiresAt,
   extend,
   logoutAction
@@ -20,35 +21,13 @@ const SessionEndingAlert = ({
     ? 'alert--session-expiring__active'
     : 'alert--session-expiring__inactive';
 
-  // Renderer callback with condition
-  // eslint-disable-next-line react/prop-types
-  const createTimer = ({ minutes, seconds, completed }) => {
-    if (completed) {
-      // Render a completed state
-      return (
-        <Alert
-          heading="Your session has ended."
-          role="alertdialog"
-          variation="error"
-        />
-      );
-    }
-    // Render a countdown
+  /* eslint-disable react/prop-types */
+  const createTimer = ({ minutes, seconds }) => {
     return (
-      <Alert
-        heading="Your session is expiring."
-        role="alertdialog"
-        variation="warn"
-      >
+      <span>
         Your session will end in {zeroPad(minutes)}:{zeroPad(seconds)} minutes.
         If you’d like to keep working, choose continue.
-        <p className="ds-u-text-align--right ds-u-margin-bottom--0">
-          <Button variation="primary" onClick={extend}>
-            {isExtendingSession && <Spinner />}
-            {isExtendingSession ? ' Continuing' : 'Continue'}
-          </Button>
-        </p>
-      </Alert>
+      </span>
     );
   };
 
@@ -58,13 +37,45 @@ const SessionEndingAlert = ({
       aria-live="polite"
       className={`alert--session-expiring ${className}`}
     >
+      <div id="session-ending-modal" />
       {isSessionEnding && (
-        <Countdown
-          date={expiresAt - 5000}
+        <Dialog
           key={uuidv4()}
-          renderer={createTimer}
-          onComplete={logoutAction}
-        />
+          heading="Your session is expiring."
+          alert
+          getApplicationNode={() =>
+            document.getElementById('session-ending-modal')
+          }
+          actionsClassName="ds-u-text-align--right ds-u-margin-bottom--0"
+          actions={[
+            <Button variation="primary" onClick={extend} key={uuidv4()}>
+              {isExtendingSession && <Spinner />}
+              {isExtendingSession ? ' Continuing' : 'Continue'}
+            </Button>,
+            <Button
+              variation="transparent"
+              onClick={logoutAction}
+              key={uuidv4()}
+            >
+              {isLoggingOut && <Spinner />}
+              {isLoggingOut ? ' Signing out' : 'Sign out'}
+            </Button>
+          ]}
+          ariaCloseLabel="Close and continue"
+          closeText="Close"
+          onExit={extend}
+          underlayClickExits={false}
+          escapeExitDisabled
+          size="wide"
+        >
+          <Countdown
+            date={expiresAt - 5000}
+            key={uuidv4()}
+            renderer={createTimer}
+            onComplete={logoutAction}
+            onStop={extend}
+          />
+        </Dialog>
       )}
     </div>
   );
@@ -73,20 +84,25 @@ const SessionEndingAlert = ({
 SessionEndingAlert.propTypes = {
   isSessionEnding: PropTypes.bool.isRequired,
   isExtendingSession: PropTypes.bool.isRequired,
+  isLoggingOut: PropTypes.bool.isRequired,
   expiresAt: PropTypes.number.isRequired,
   extend: PropTypes.func.isRequired,
   logoutAction: PropTypes.func.isRequired
 };
 
 const mapStateToProps = ({
-  auth: { isSessionEnding, isExtendingSession, expiresAt }
+  auth: { isSessionEnding, isExtendingSession, isLoggingOut, expiresAt }
 }) => ({
   isSessionEnding,
   isExtendingSession,
+  isLoggingOut,
   expiresAt
 });
 
-const mapDispatchToProps = { extend: extendSession, logoutAction: logout };
+const mapDispatchToProps = {
+  extend: extendSession,
+  logoutAction: logout
+};
 
 export default connect(mapStateToProps, mapDispatchToProps)(SessionEndingAlert);