+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Anticipated release: December 21, 2020
 
 - fixed security headers
 - displays images within the tinymce editor ([#2348])
+- fixed session expiring warning bug ([#2720])
 
 #### ⚙️ Behind the scenes
 
@@ -23,11 +24,11 @@ Anticipated release: December 21, 2020
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
-
 [#2348]: https://github.com/CMSgov/eAPD/issues/2348
 [#2682]: https://github.com/CMSgov/eAPD/issues/2682
 [#2583]: https://github.com/CMSgov/eAPD/issues/2583
 [#2692]: https://github.com/CMSgov/eAPD/issues/2692
 [#2702]: https://github.com/CMSgov/eAPD/issues/2702
+[#2720]: https://github.com/CMSgov/eAPD/issues/2720
 [ghsa-vrv8-v4w8-f95h]: https://github.com/advisories/GHSA-vrv8-v4w8-f95h
 [ghsa-w7rc-rwvf-8q5r]: https://github.com/advisories/GHSA-w7rc-rwvf-8q5r